Поправлен тип данных, реализован метод сложения(plus)

+++ src/main/java/com/example/calculatorskypro/CalculatorController.java
@@ -20,7 +20,7 @@ public class CalculatorController {
     }
 
     @GetMapping(path = "/plus")
-    public String calculatorPlus(@RequestParam("num1") float num1, @RequestParam("num2") float num2){
+    public String calculatorPlus(@RequestParam("num1") Integer num1, @RequestParam("num2") Integer num2){
         return calculatorService.calculatorPlus(num1, num2);
     }
 }
